Chatbot_langchain_vectordb_openai/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.1
Deprecated==1.2.14
effdet==0.4.1
emoji==2.8.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
filetype==1.2.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.0
grpcio==1.59.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.330
langdetect==1.0.9
langsmith==0.0.57
layoutparser==0.3.4
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.1
opencv-python==4.8.1.78
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
playwright==1.39.0
portalocker==2.8.2
posthog==3.0.2
protobuf==4.25.0
pulsar-client==3.3.0
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.16.1
pyparsing==3.1.1
pypdfium2==4.23.1
PyPika==0.48.9
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
rapidfuzz==3.5.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
safetensors==0.4.0
scipy==1.11.3
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.28.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
timm==0.9.10
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
unstructured==0.10.28
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0